requirements.txt
matplotlib==3.6.2
nltk==3.8.1
numpy==1.23.3
Pillow==9.4.0
psycopg2==2.9.5
pyTelegramBotAPI==4.9.0
requests==2.28.1
schedule==1.1.0
torch==1.13.1
